Refactor input system to use double precision values
Replaces most float and Float2 usages with double and Double2 in input management interfaces, processors, actions, and composites. This improves precision for input values such as mouse deltas, gamepad axes, triggers, and timing parameters. Updates method signatures, property types, and internal calculations accordingly.

+++ Prowl.Runtime/InputManagement/InputBinding.cs
@@ -72,7 +72,7 @@ public class InputBinding
     /// <summary>
     /// Duration (in seconds) required for Hold interaction. Default: 0.4s
     /// </summary>
-    public float HoldDuration { get; set; } = 0.4f;
+    public double HoldDuration { get; set; } = 0.4f;
 
     /// <summary>
     /// Number of taps required for MultiTap interaction. Default: 2
@@ -82,12 +82,12 @@ public class InputBinding
     /// <summary>
     /// Maximum time window (in seconds) between taps for MultiTap. Default: 0.3s
     /// </summary>
-    public float TapWindow { get; set; } = 0.3f;
+    public double TapWindow { get; set; } = 0.3f;
 
     /// <summary>
     /// Maximum duration (in seconds) for a Tap to be valid. Default: 0.2s
     /// </summary>
-    public float MaxTapDuration { get; set; } = 0.2f;
+    public double MaxTapDuration { get; set; } = 0.2f;
 
     /// <summary>
     /// Optional processors to apply to the input value (e.g., normalize, clamp, invert).
